django/source/requirements.txt
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
confusable-homoglyphs==3.2.0
Django==3.1
django-appconf==1.0.4
django-compressor==2.4
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-libsass==0.8
django-mptt==0.11.0
django-registration==3.1
django-storages==1.10.1
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.10
libsass==0.20.0
Pillow==7.2.0
protobuf==3.13.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pytz==2020.1
rcssmin==1.0.6
requests==2.24.0
rjsmin==1.1.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10